refactor: move utility functions to count pages in PDF files to utils folder
